change files structure, rename dumb components
