+++ client/src/utils/formatDate.ts
@@ -1,7 +1,7 @@
 import { format } from "date-fns";
 
 export const formatDate = (time?: string | number | Date) => {
-    if  (!time) {
+    if (!time) {
         return;
     }
